refactor app into multiple files
